Controversy at the PTO
This is a paper about the arguing that Patent Acts create a kind of
revolution defining and shaping technology for a specific time and
place. It is written in LaTeX

The United States Constitution instructed the First Congress of the
United States to enact laws "securing for limited Times to Authors and
Inventors the exclusive Right to their respective Writings and
Discoveries." This instruction, commonly referred to as the
Intellectual Property Clause, demanded a set of laws based on the broad
idea that inventors are best incentivized and remunerated for their
works through a system of guaranteed exclusive rights. What exactly the
scope of those rights were, who had authority to hand them out, and
under what conditions they would be given, were all left to Congress to
sort out; it was among the earliest controversies over science and
technology in the United States of America.